Revert "Plan 9 from Bell Labs Fourth Edition 2004-08-06"
This reverts commit 061a3f44e70ec02be7c1ae5efce31e69be4bb7e7.

+++ sys/src/cmd/5i/5i.c
@@ -6,7 +6,7 @@
 
 char*	file = "5.out";
 int	datasize;
-ulong	textbase;
+int	textbase;
 Biobuf	bp, bi;
 Fhdr	fhdr;
