+++ sys/src/cmd/vi/vi.c
@@ -7,7 +7,7 @@
 
 char	*file = "v.out";
 int	datasize;
-ulong	textbase;
+int	textbase;
 Biobuf	bp, bi;
 Fhdr	fhdr;
